requirements.txt
fastapi
uvicorn
requests
pillow
python-dotenv 
pydantic